+++ test/cli/output/test_default.py
@@ -159,7 +159,7 @@ def test_display_single_failure(capsys, swagger_20, endpoint, body):
         endpoint, [success, success, success, failure, failure, models.Check("different_check", models.Status.success)]
     )
     # When this failure is displayed
-    default.display_single_failure(test_statistic)
+    default.display_failures_for_single_test(test_statistic)
     out = capsys.readouterr().out
     lines = out.split("\n")
     # Then the endpoint name is displayed as a subsection